+++ Final_Fantasy_Mystery_World/m1EasingSplines.h
@@ -25,7 +25,7 @@ struct EaseFunctions {
 };
 
 struct EaseSplineInfo {
-	int * position = nullptr;
+	float * position = nullptr;
 	TypeSpline type;
 	int initial_position;
 	int distance_to_travel;
@@ -34,7 +34,7 @@ struct EaseSplineInfo {
 	EaseFunctions ease_function;
 	bool Update(float dt);
 
-	EaseSplineInfo(int * position, const int target_position, const float time_to_travel, TypeSpline type) {
+	EaseSplineInfo(float * position, const int target_position, const float time_to_travel, TypeSpline type) {
 		this->position = position;
 		this->initial_position = *position;
 		this->distance_to_travel = target_position - *position;
@@ -60,7 +60,7 @@ public:
 	// Called before quitting
 	bool CleanUp();
 
-	void CreateSpline(int * position, int target_position, const float time_to_travel, TypeSpline type);
+	void CreateSpline(float * position, int target_position, const float time_to_travel, TypeSpline type);
 
 private:
